Progress on abstraction, needs work for save
