+++ lib/draft/PickDataFetchers/parsing_helpers.ex
